+++ src/main/kotlin/dev/mizarc/waystonewarps/menus/WarpManagementMenu.kt
@@ -10,7 +10,6 @@ import com.github.stefvanschie.inventoryframework.gui.type.HopperGui
 import com.github.stefvanschie.inventoryframework.pane.StaticPane
 import com.github.stefvanschie.inventoryframework.pane.component.Label
 import dev.mizarc.waystonewarps.*
-import dev.mizarc.waystonewarps.domain.waystones.WaystoneRepositorySQLite
 import dev.mizarc.waystonewarps.utils.getWarpMoveTool
 import org.bukkit.Bukkit
 import org.bukkit.Material
